+++ superduperdb/ext/anthropic/__init__.py
@@ -1,3 +1,7 @@
+from superduperdb.misc.annotations import requires_packages
+
+requires_packages(['anthropic', '0.12.0'])
+
 from .model import AnthropicCompletions
 
 __all__ = ('AnthropicCompletions',)