+++ superduperdb/ext/unstructured/__init__.py
@@ -1,0 +1,3 @@
+from superduperdb.misc.annotations import requires_packages
+
+requires_packages(['unstructured', '0.12.5'])